// src/types/enums.ts
// todo: enums EInvoice

export enum InvoiceStatus {
    DRAFT = 'DRAFT',
    APPROVAL_PENDING = 'APPROVAL_PENDING',
    ACCEPTED = 'ACCEPTED',
    REJECTED = 'REJECTED',
    PAID = 'PAID',
    OVERDUE = 'OVERDUE',
    QUICKPAY_AVAILABLE = 'QUICKPAY_AVAILABLE',
    QUICKPAY_AVAILABLE_TEXT = 'QUICKPAY AVAILABLE',
    QUICKPAY = 'QUICKPAY',
    IN_PROGRESS = 'IN_PROGRESS',
	ADVANCED = 'ADVANCED',
	ADVANCE = 'ADVANCE',
	DEFEATED = 'DEFEATED',
    PREPAYMENT = 'PREPAYMENT'
}


export enum TypeCompany {
	SOLE_PROPRIETORSHIP = 'SOLE_PROPRIETORSHIP',
	CORPORATION = 'CORPORATION',

}
export enum NotificationStatus {
    SUPPLIER_UPLOADS_INVOICE = 'SUPPLIER_UPLOADS_INVOICE',
	SUPPLIER_ACCEPTS_QUICKPAY_OFFER = 'SUPPLIER_ACCEPTS_QUICKPAY_OFFER',
	INVOICE_DUE_DAY = 'INVOICE_DUE_DAY',
	PAYER_REJECTS_INVOICE = 'PAYER_REJECTS_INVOICE',
	PAYER_ACCEPTS_INVOICE_WITHOUT_QUICKPAY = 'PAYER_ACCEPTS_INVOICE_WITHOUT_QUICKPAY',
	PAYER_ACCEPTS_INVOICE_WITH_QUICKPAY_OFFER = 'PAYER_ACCEPTS_INVOICE_WITH_QUICKPAY_OFFER',
	INVOICE_MARKED_AS_PAID_OR_ADVANCED = 'INVOICE_MARKED_AS_PAID_OR_ADVANCED',
	INVOICE_EXTERNAL = 'INVOICE_EXTERNAL',
	PAY_PAYER_TO_FINANCIAL = 'PAY_PAYER_TO_FINANCIAL'
}

export enum IInvoicesType {
    RECEIVED = 'received',
    SENT = 'sent'
}

export enum EOnboardProfileStatus {
    NEW = 'NEW',
    APPROVAL_PENDING = 'APPROVAL_PENDING',
    ACCEPTED = 'ACCEPTED',
    REJECTED = 'REJECTED'
}

export enum EUserType {
    SUPPLIER = 'SUPPLIER',
    PAYER = 'PAYER',
    BOTH = 'BOTH',
    FINANCIAL = 'FINANCIAL',
}


export enum EUserTypeE {
    DISTRIBUTOR = 'DISTRIBUTOR',
    BUILDER = 'BUILDER',
    WHOLESALER = 'WHOLESALER',
    SENIOR_WHOLESALER = 'SENIOR_WHOLESALER',
    SILVER_PRODUCER = 'SILVER_PRODUCER',
    GOLDEN_PRESIDENT = 'GOLDEN_PRESIDENT'
}